backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [{ name = "SherlockH0", email = "ddmlf7@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
  "django (>=5.2,<6.0)",
  "django-split-settings (>=1.3.2,<2.0.0)",
  "pyyaml (>=6.0.2,<7.0.0)",
  "djangorestframework (>=3.16.0,<4.0.0)",
  "pillow (>=11.2.1,<12.0.0)",
  "daphne (>=4.1.2,<5.0.0)",
  "channels[daphne] (>=4.2.2,<5.0.0)",
  "channels-redis (>=4.2.1,<5.0.0)",
  "django-filter (>=25.1,<26.0)",
  "bleach (>=6.2.0,<7.0.0)",
  "celery (>=5.5.2,<6.0.0)",
  "lxml (>=5.4.0,<6.0.0)",
  "redis[hiredis] (>=6.0.0,<7.0.0)",
  "django-redis (>=5.4.0,<6.0.0)",
  "python-magic (>=0.4.27,<0.5.0)",
  "psycopg (>=3.2.7,<4.0.0)",
  "twisted[http2,tls] (>=24.11.0,<25.0.0)",
  "django-simple-captcha (>=0.6.2,<0.7.0)",
  "pytidylib (>=0.3.2,<0.4.0)",
  "django-cors-headers (>=4.7.0,<5.0.0)",
  "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
django-types = "^0.20.0"
djangorestframework-types = "^0.9.0"
celery-types = "^0.23.0"

# [tool.poetry]
# package-mode = false
